[refactor] create new use case in account and his test

+++ routes/Role/Role.php
@@ -5,8 +5,8 @@ use Src\Role\Infrastructure\Controllers\RoleGetController;
 use Src\Role\Infrastructure\Controllers\RolePostController;
 use Src\Role\Infrastructure\Controllers\RolePutController;
 
-Route::get('/read', [RoleGetController::class, 'read']);
-Route::get('/{id}/show', [RoleGetController::class, 'show']);
-Route::post('/create', [RolePostController::class, 'create']);
-Route::put('/{id}/update', [RolePutController::class, 'update']);
-Route::delete('/{id}/delete', [RoleDeleteController::class, 'delete']);
+Route::middleware(['scope:admin'])->get('/read', [RoleGetController::class, 'read']);
+Route::middleware(['scope:admin,guest'])->get('/{id}/show', [RoleGetController::class, 'show']);
+Route::middleware(['scope:admin'])->post('/create', [RolePostController::class, 'create']);
+Route::middleware(['scope:admin'])->put('/{id}/update', [RolePutController::class, 'update']);
+Route::middleware(['scope:admin'])->delete('/{id}/delete', [RoleDeleteController::class, 'delete']);
